refactor(gql): rename gql_schema to graphql and convert to Django app
Rename package from gql_schema to graphql for consistency with project
naming conventions (api, alerts, gtfs, etc.) and convert to proper Django app.

Changes:
- Rename gql_schema/ → graphql/ directory
- Create graphql/apps.py with GraphqlConfig class
- Update graphql/__init__.py with app config and documentation
- Add graphql.apps.GraphqlConfig to INSTALLED_APPS in settings
- Update all imports (datahub/urls.py, demo_graphql.py)

Benefits:
- Consistent with project naming (no underscores, no abbreviations)
- Proper Django app structure (recognized by manage.py commands)
- Better documentation and discoverability
- Can have migrations/models/management commands if needed later
- More professional Django convention

Addresses feedback to make GraphQL its own app.
