+++ datahub/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     "feed.apps.FeedConfig",
     "alerts.apps.AlertsConfig",
     "api.apps.ApiConfig",
+    "graphql.apps.GraphqlConfig",
     "rest_framework",
     "rest_framework.authtoken",
     "drf_spectacular",
