--- gql_schema/__init__.py
@@ -1,1 +1,0 @@
-# GraphQL package for Infobús project
